pyproject.toml
[project]
name = "chktrade"
version = "0.1.0"
description = "All in one trade platform"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "textual>=2.1.0",
]

[dependency-groups]
dev = [
    "textual-dev>=1.7.0",
]
